+++ hack/update-checksum-windows.sh
@@ -21,15 +21,12 @@ YAMLS_DIR="${WORK_DIR}"/../build/yamls
 MANIFESTS=$(ls $YAMLS_DIR/antrea-windows*.yml)
 WINDOWS_DIR="${YAMLS_DIR}"/windows
 BASE_CONF_FILES="${WINDOWS_DIR}/base/conf/antrea-agent.conf ${WINDOWS_DIR}/base/conf/antrea-cni.conflist"
-DEFAULT_CONF_FILES="${WINDOWS_DIR}/default/conf/Run-AntreaAgent.ps1"
 CONTAINERD_CONF_FILES="${WINDOWS_DIR}/containerd/conf/Install-WindowsCNI-Containerd.ps1 \
     ${WINDOWS_DIR}/containerd/conf/Run-AntreaAgent-Containerd.ps1"
 CONTAINERD_WITH_OVS_CONF_FILES="${WINDOWS_DIR}/containerd-with-ovs/conf/Run-AntreaOVS-Containerd.ps1 \
     ${WINDOWS_DIR}/containerd-with-ovs/conf/VMSwitchExtension-AntreaAgent-Containerd.ps1"
 
 checksum_windows_config=$(cat ${BASE_CONF_FILES} | sha256sum | cut -d " " -f 1)
-
-checksum_default=$(cat ${DEFAULT_CONF_FILES} | sha256sum | cut -d " " -f 1)
 
 checksum_containerd=$( cat ${CONTAINERD_CONF_FILES} | sha256sum | cut -d " " -f 1)
 
@@ -39,6 +36,5 @@ for file in ${MANIFESTS[@]}; do
     sed -i.bak "s/windows-config-checksum-placeholder/${checksum_windows_config}/g" ${file}
 done
 
-sed -i.bak "s/agent-windows-checksum-placeholder/${checksum_default}/g" ${YAMLS_DIR}/antrea-windows.yml
 sed -i.bak "s/agent-windows-checksum-placeholder/${checksum_containerd}/g" ${YAMLS_DIR}/antrea-windows-containerd.yml
 sed -i.bak "s/agent-windows-checksum-placeholder/${checksum_containerd_with_ovs}/g" ${YAMLS_DIR}/antrea-windows-containerd-with-ovs.yml
