+++ pkg/agent/cniserver/pod_configuration.go
@@ -55,9 +55,6 @@ const (
 )
 
 const (
-	defaultOVSInterfaceType int = iota //nolint suppress deadcode check for windows
-	internalOVSInterfaceType
-
 	defaultIFDevName = "eth0"
 )
 
@@ -265,15 +262,11 @@ func (pc *podConfigurator) configureInterfacesCommon(
 func (pc *podConfigurator) createOVSPort(ovsPortName string, ovsAttachInfo map[string]interface{}, vlanID uint16) (string, error) {
 	var portUUID string
 	var err error
-	switch getOVSInterfaceType(ovsPortName) {
-	case internalOVSInterfaceType:
-		portUUID, err = pc.ovsBridgeClient.CreateInternalPort(ovsPortName, 0, "", ovsAttachInfo)
-	default:
-		if vlanID == 0 {
-			portUUID, err = pc.ovsBridgeClient.CreatePort(ovsPortName, ovsPortName, ovsAttachInfo)
-		} else {
-			portUUID, err = pc.ovsBridgeClient.CreateAccessPort(ovsPortName, ovsPortName, ovsAttachInfo, vlanID)
-		}
+
+	if vlanID == 0 {
+		portUUID, err = pc.ovsBridgeClient.CreatePort(ovsPortName, ovsPortName, ovsAttachInfo)
+	} else {
+		portUUID, err = pc.ovsBridgeClient.CreateAccessPort(ovsPortName, ovsPortName, ovsAttachInfo, vlanID)
 	}
 	if err != nil {
 		klog.Errorf("Failed to add OVS port %s, remove from local cache: %v", ovsPortName, err)
@@ -490,55 +483,6 @@ func (pc *podConfigurator) reconcile(pods []corev1.Pod, containerAccess *contain
 		klog.ErrorS(err, "Error when garbage collecting previously-allocated IPs")
 	}
 
-	return nil
-}
-
-func (pc *podConfigurator) connectInterfaceToOVSCommon(ovsPortName, netNS string, containerConfig *interfacestore.InterfaceConfig) error {
-	// create OVS Port and add attach container configuration into external_ids
-	containerID := containerConfig.ContainerID
-	klog.V(2).Infof("Adding OVS port %s for container %s", ovsPortName, containerID)
-	ovsAttachInfo := BuildOVSPortExternalIDs(containerConfig)
-	portUUID, err := pc.createOVSPort(ovsPortName, ovsAttachInfo, containerConfig.VLANID)
-	if err != nil {
-		return fmt.Errorf("failed to add OVS port for container %s: %v", containerID, err)
-	}
-	// Remove OVS port if any failure occurs in later manipulation.
-	defer func() {
-		if err != nil {
-			_ = pc.ovsBridgeClient.DeletePort(portUUID)
-		}
-	}()
-
-	var ofPort int32
-	// Not needed for a secondary network interface.
-	if !pc.isSecondaryNetwork {
-		// GetOFPort will wait for up to 1 second for OVSDB to report the OFPort number.
-		ofPort, err = pc.ovsBridgeClient.GetOFPort(ovsPortName, false)
-		if err != nil {
-			return fmt.Errorf("failed to get of_port of OVS port %s: %v", ovsPortName, err)
-		}
-		klog.V(2).InfoS("Setting up Openflow entries for Pod interface", "container", containerID, "port", ovsPortName)
-		if err = pc.ofClient.InstallPodFlows(ovsPortName, containerConfig.IPs, containerConfig.MAC, uint32(ofPort), containerConfig.VLANID, nil); err != nil {
-			return fmt.Errorf("failed to add Openflow entries for container %s: %v", containerID, err)
-		}
-	}
-
-	containerConfig.OVSPortConfig = &interfacestore.OVSPortConfig{PortUUID: portUUID, OFPort: ofPort}
-	// Add containerConfig into local cache
-	pc.ifaceStore.AddInterface(containerConfig)
-
-	// Not needed for a secondary network interface.
-	if !pc.isSecondaryNetwork {
-		// Notify the Pod update event to required components.
-		event := agenttypes.PodUpdate{
-			PodName:      containerConfig.PodName,
-			PodNamespace: containerConfig.PodNamespace,
-			ContainerID:  containerConfig.ContainerID,
-			NetNS:        netNS,
-			IsAdd:        true,
-		}
-		pc.podUpdateNotifier.Notify(event)
-	}
 	return nil
 }
